added tests for QuoteController
