--- tests/Controller/QuoteControllerTest.php
@@ -1,129 +1,0 @@
-<?php
-
-namespace App\Test\Controller;
-
-use App\Entity\Quote;
-use App\Repository\QuoteRepository;
-use Symfony\Bundle\FrameworkBundle\KernelBrowser;
-use Symfony\Bundle\FrameworkBundle\Test\WebTestCase;
-
-class QuoteControllerTest extends WebTestCase
-{
-    private KernelBrowser $client;
-    private QuoteRepository $repository;
-    private string $path = '/admin/quote/';
-
-    protected function setUp(): void
-    {
-        $this->client = static::createClient();
-        $this->repository = static::getContainer()->get('doctrine')->getRepository(Quote::class);
-
-        foreach ($this->repository->findAll() as $object) {
-            $this->repository->remove($object, true);
-        }
-    }
-
-    public function testIndex(): void
-    {
-        $crawler = $this->client->request('GET', $this->path);
-
-        self::assertResponseStatusCodeSame(200);
-        self::assertPageTitleContains('Quote index');
-
-        // Use the $crawler to perform additional assertions e.g.
-        // self::assertSame('Some text on the page', $crawler->filter('.p')->first());
-    }
-
-    public function testNew(): void
-    {
-        $originalNumObjectsInRepository = count($this->repository->findAll());
-
-        $this->markTestIncomplete();
-        $this->client->request('GET', sprintf('%snew', $this->path));
-
-        self::assertResponseStatusCodeSame(200);
-
-        $this->client->submitForm('Save', [
-            'quote[price]' => 'Testing',
-            'quote[title]' => 'Testing',
-            'quote[user]' => 'Testing',
-            'quote[recipes]' => 'Testing',
-        ]);
-
-        self::assertResponseRedirects('/admin/quote/');
-
-        self::assertSame($originalNumObjectsInRepository + 1, count($this->repository->findAll()));
-    }
-
-    public function testShow(): void
-    {
-        $this->markTestIncomplete();
-        $fixture = new Quote();
-        $fixture->setPrice('My Title');
-        $fixture->setTitle('My Title');
-        $fixture->setUser('My Title');
-        $fixture->setRecipes('My Title');
-
-        $this->repository->save($fixture, true);
-
-        $this->client->request('GET', sprintf('%s%s', $this->path, $fixture->getId()));
-
-        self::assertResponseStatusCodeSame(200);
-        self::assertPageTitleContains('Quote');
-
-        // Use assertions to check that the properties are properly displayed.
-    }
-
-    public function testEdit(): void
-    {
-        $this->markTestIncomplete();
-        $fixture = new Quote();
-        $fixture->setPrice('My Title');
-        $fixture->setTitle('My Title');
-        $fixture->setUser('My Title');
-        $fixture->setRecipes('My Title');
-
-        $this->repository->save($fixture, true);
-
-        $this->client->request('GET', sprintf('%s%s/edit', $this->path, $fixture->getId()));
-
-        $this->client->submitForm('Update', [
-            'quote[price]' => 'Something New',
-            'quote[title]' => 'Something New',
-            'quote[user]' => 'Something New',
-            'quote[recipes]' => 'Something New',
-        ]);
-
-        self::assertResponseRedirects('/admin/quote/');
-
-        $fixture = $this->repository->findAll();
-
-        self::assertSame('Something New', $fixture[0]->getPrice());
-        self::assertSame('Something New', $fixture[0]->getTitle());
-        self::assertSame('Something New', $fixture[0]->getUser());
-        self::assertSame('Something New', $fixture[0]->getRecipes());
-    }
-
-    public function testRemove(): void
-    {
-        $this->markTestIncomplete();
-
-        $originalNumObjectsInRepository = count($this->repository->findAll());
-
-        $fixture = new Quote();
-        $fixture->setPrice('My Title');
-        $fixture->setTitle('My Title');
-        $fixture->setUser('My Title');
-        $fixture->setRecipes('My Title');
-
-        $this->repository->save($fixture, true);
-
-        self::assertSame($originalNumObjectsInRepository + 1, count($this->repository->findAll()));
-
-        $this->client->request('GET', sprintf('%s%s', $this->path, $fixture->getId()));
-        $this->client->submitForm('Delete');
-
-        self::assertSame($originalNumObjectsInRepository, count($this->repository->findAll()));
-        self::assertResponseRedirects('/admin/quote/');
-    }
-}
